rag-msmarco-vllm/requirements.txt
# Core LangChain dependencies
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.0.5
langchain-huggingface>=0.0.1
langchain-text-splitters>=0.0.1
langchain-community>=0.0.10

# Data processing and ML
datasets>=2.14.0
huggingface_hub>=0.17.0
sentence-transformers>=5.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Vector database
qdrant-client>=1.6.0

# API and web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.0
httpx>=0.25.0

# Monitoring and observability
prometheus-client>=0.17.0
structlog>=23.1.0

# Utilities
python-ulid>=1.1.0
orjson>=3.9.0
python-multipart>=0.0.6

# Optional: OpenTelemetry for tracing
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-exporter-otlp>=1.20.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0  # for async HTTP client testing
aiohttp>=3.8.0  # for load testing